% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/national_library_medicine.R
\name{searchnlm}
\alias{searchnlm}
\title{Takes a search term and queries the National Library of Medicine}
\usage{
searchnlm(term, field = NA, retmax = NA, email = NA)
}
\arguments{
\item{term}{The search term to query. By default spaces will be parsed as ANDs, and seperate elements in the vector as ORs.}

\item{field}{The field to query. Defaults to querying all fields. Includes choices such as "creator", "subject", "title", and "description"}

\item{retmax}{The number of observations you want to query. Defaults to 10.}

\item{email}{Your email address as a string. Optional--it allows the National Library of Medicine to contact you if there are problems with your queries.}
}
\value{
Returns a dataframe where each row is a field of a record in the NLM databse. Rank and url uniquely identify the works in the NLM database.
}
\description{
Takes a search term and queries the National Library of Medicine
}
