% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidynlm.R
\docType{package}
\name{tidynlm}
\alias{tidynlm}
\title{tidynlm: An API Wrapper for the National Library of Medicine}
\description{
tidynlm provides access to the National Library of Medicine (NLM) database.
Queries are executed using the \code{\link{searchnlm}} function. Individual
documents can be downloaded using the URLs returned by searchnlm.
}
